Switch from Mutex to RefCell for app_state
Because there are no threads.
Reentrant window proc is still a danger, so this is an attempt
to build sort of safe sort of abstraction around it.

The idea is that WinAPI calls that can dispatch messages
should be wrapped in functions that take &mut RefCell<AppState>.
This ensures that it's not borrowed at the moment and
on reentering window proc we'll be able to borrow app state.